refactor: lib 지우기
폴더명이 길어지기 때문에
