--- src/liblogger/logger.hpp
@@ -1,8 +1,0 @@
-#ifndef LIBLOGGER_LOGGER_HPP
-#define LIBLOGGER_LOGGER_HPP
-
-
-namespace logger {
-}  // namespace logger
-
-#endif // LIBLOGGER_LOGGER_HPP
