+++ src/logger/logger.hpp
@@ -1,0 +1,6 @@
+#ifndef LOGGER_LOGGER_HPP
+#define LOGGER_LOGGER_HPP
+
+namespace logger {}  // namespace logger
+
+#endif  // LOGGER_LOGGER_HPP
