+++ src/strutil/strutil.cpp
