+++ src/strutil/strutil.hpp
